--- BEA/labs/lab1.txt
@@ -1,2 +1,0 @@
-Вставляем первую половину текста
-Добавляем вторую половину текста
