--- BEA/labs/testGit.txt
@@ -1,4 +1,0 @@
-first commit
-second commit
-third commit
-fourth commit
