+++ Lab7/src/Command.java
@@ -1,0 +1,5 @@
+// Интерфейс команды
+interface Command {
+    void execute();
+    void undo();
+}
